[REFACTOR] separate public redirect and project view from CRUD controllers, add IP normalization middleware, and improve IpInfoService for real user tracking
