+++ app/Http/Controllers/ProjectController.php
@@ -256,92 +256,4 @@ class ProjectController extends Controller
 
         return redirect()->route('projects.index')->with('success', 'Selected Deleted Successfully.');
     }
-
-    /**
-     * Menampilkan halaman publik berdasarkan slug project.
-     */
-    public function showBySlug(Request $request, $slug)
-    {
-        $project = Project::where('project_slug', $slug)->firstOrFail();
-
-        // Jika project punya password
-        if ($project->password) {
-            if (!$request->session()->has("project_access_{$project->id}")) {
-                return Inertia::render('PasswordForm', [
-                    'project_slug' => $project->project_slug,
-                ]);
-            }
-        }
-
-        ProjectClick::create([
-            'project_id' => $project->id,
-            'ip_address' => $request->ip(),
-            'user_agent' => $request->userAgent(),
-            'referer'    => $request->headers->get('referer'),
-            'country'    => null, // bisa isi geoip
-        ]);
-
-        $mainLinks = $project->links()
-            ->whereNull('parent_id')
-            ->where('is_active', true)
-            ->with([
-                'children' => function ($query) {
-                    $query->where('is_active', true);
-                },
-                'category'
-            ])
-            ->get();
-
-        $grouped = $mainLinks->groupBy(function ($link) {
-            return optional($link->category)->name ?? 'Uncategorized';
-        })->map(function ($links, $categoryName) {
-            $categoryImage = optional($links->first()->category)->image_url;
-            return [
-                'category' => $categoryName,
-                'image_url' => $categoryImage,
-                'links' => $links->map(function ($link) {
-                    return [
-                        'id' => $link->id,
-                        'title' => $link->title,
-                        'original_url' => $link->original_url,
-                        'children' => $link->children->map(function ($child) {
-                            return [
-                                'title' => $child->title,
-                                'original_url' => $child->original_url,
-                            ];
-                        })->values()
-                    ];
-                })->values()
-            ];
-        })->values();
-
-        return Inertia::render('Projects/PublicView', [
-            'project' => [
-                'id' => $project->id,
-                'name' => $project->project_name,
-                'slug' => $project->project_slug,
-            ],
-            'categories' => $grouped,
-        ]);
-    }
-
-    public function verifyPassword(Request $request, $slug)
-    {
-        $project = Project::where('project_slug', $slug)->firstOrFail();
-
-        $request->validate([
-            'password' => 'required|string',
-        ]);
-
-        if (!\Hash::check($request->password, $project->password)) {
-            return response()->json(['error' => 'Password is incorrect.'], 422);
-        }
-
-        $request->session()->put("project_access_{$project->id}", true);
-
-        return response()->json([
-            'success' => true,
-            'url' => route('projects.showBySlug', $project->project_slug),
-        ]);
-    }
 }
